ml_pipeline/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
geopandas>=0.14.0
rasterio>=1.3.0
xarray>=2023.1.0
netCDF4>=1.6.0
requests>=2.31.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
shapely>=2.0.0
pyproj>=3.6.0
scipy>=1.11.0
supabase>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
